control

+++ robot_ws/src/nsra_controller/nsra_control/include/nsra_control/nsra_hw_interface.h
@@ -85,7 +85,7 @@ public:
 
   serial::Serial ser;
 
-  int test_pos = 0;
+  uint8_t test_pos = 0;
 
   //LibSerial::SerialStream serial_stream;
